+++ src/main/java/learn/thyme/thyteams/user/UserServiceImpl.java
@@ -90,4 +90,9 @@ public class UserServiceImpl implements UserService {
     public void deleteAllUsers() {
         repository.deleteAll();
     }
+
+    @Override
+    public long countUsers() {
+        return repository.count();
+    }
 }